requirements.txt
pytest
pytest-bdd
pytest-playwright
playwright
pyyaml